requirements.txt
customtkinter
tinydb
babel
PyYAML